# Server-Python/swagger_server/models/gudang_komoditas.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class GudangKomoditas(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, gudang_id: int=None, nama: str=None, kategori: str=None, lokasi: str=None):  # noqa: E501
        """GudangKomoditas - a model defined in Swagger

        :param gudang_id: The gudang_id of this GudangKomoditas.  # noqa: E501
        :type gudang_id: int
        :param nama: The nama of this GudangKomoditas.  # noqa: E501
        :type nama: str
        :param kategori: The kategori of this GudangKomoditas.  # noqa: E501
        :type kategori: str
        :param lokasi: The lokasi of this GudangKomoditas.  # noqa: E501
        :type lokasi: str
        """
        self.swagger_types = {
            'gudang_id': int,
            'nama': str,
            'kategori': str,
            'lokasi': str
        }

        self.attribute_map = {
            'gudang_id': 'gudang_id',
            'nama': 'Nama',
            'kategori': 'Kategori',
            'lokasi': 'Lokasi'
        }
        self._gudang_id = gudang_id
        self._nama = nama
        self._kategori = kategori
        self._lokasi = lokasi

    @classmethod
    def from_dict(cls, dikt) -> 'GudangKomoditas':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The GudangKomoditas of this GudangKomoditas.  # noqa: E501
        :rtype: GudangKomoditas
        """
        return util.deserialize_model(dikt, cls)

    @property
    def gudang_id(self) -> int:
        """Gets the gudang_id of this GudangKomoditas.


        :return: The gudang_id of this GudangKomoditas.
        :rtype: int
        """
        return self._gudang_id

    @gudang_id.setter
    def gudang_id(self, gudang_id: int):
        """Sets the gudang_id of this GudangKomoditas.


        :param gudang_id: The gudang_id of this GudangKomoditas.
        :type gudang_id: int
        """

        self._gudang_id = gudang_id

    @property
    def nama(self) -> str:
        """Gets the nama of this GudangKomoditas.


        :return: The nama of this GudangKomoditas.
        :rtype: str
        """
        return self._nama

    @nama.setter
    def nama(self, nama: str):
        """Sets the nama of this GudangKomoditas.


        :param nama: The nama of this GudangKomoditas.
        :type nama: str
        """

        self._nama = nama

    @property
    def kategori(self) -> str:
        """Gets the kategori of this GudangKomoditas.


        :return: The kategori of this GudangKomoditas.
        :rtype: str
        """
        return self._kategori

    @kategori.setter
    def kategori(self, kategori: str):
        """Sets the kategori of this GudangKomoditas.


        :param kategori: The kategori of this GudangKomoditas.
        :type kategori: str
        """

        self._kategori = kategori

    @property
    def lokasi(self) -> str:
        """Gets the lokasi of this GudangKomoditas.


        :return: The lokasi of this GudangKomoditas.
        :rtype: str
        """
        return self._lokasi

    @lokasi.setter
    def lokasi(self, lokasi: str):
        """Sets the lokasi of this GudangKomoditas.


        :param lokasi: The lokasi of this GudangKomoditas.
        :type lokasi: str
        """

        self._lokasi = lokasi